requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.4
pur==5.3.0
python-dateutil==2.8.1
six==1.13.0
uWSGI==2.0.18
Werkzeug==0.16.0
